pi-zero/requirements.txt
netifaces==0.10.9
adafruit_circuitpython_ina219==3.3.1
adafruit_circuitpython_bme680==3.1.3
adafruit_circuitpython_tsl2591==1.1.5
Adafruit_Blinka==4.1.0
board==1.0
paho_mqtt==1.5.0
pyserial==3.4
termcolor==1.1.0
